migrateファイル 編集5

+++ db/migrate/20220703054728_create_favorites.rb
@@ -1,7 +1,7 @@
 class CreateFavorites < ActiveRecord::Migration[6.1]
   def change
     create_table :favorites do |t|
-      t.bigint :user_id, null: false, foreign_key: true
+      t.integer :user_id, null: false, foreign_key: true
       t.integer :post_id, null: false, foreign_key: true
 
       t.timestamps
