+++ db/migrate/20220703054754_create_lists.rb
@@ -1,7 +1,7 @@
 class CreateLists < ActiveRecord::Migration[6.1]
   def change
     create_table :lists do |t|
-      t.bigint :user_id, null: false, foreign_key: true
+      t.integer :user_id, null: false, foreign_key: true
       t.integer :post_id, null: false, foreign_key: true
       t.string :memo
       t.timestamps
