+++ db/migrate/20220703054831_create_reposts.rb
@@ -1,7 +1,7 @@
 class CreateReposts < ActiveRecord::Migration[6.1]
   def change
     create_table :reposts do |t|
-      t.bigint :user_id, null: false, foreign_key: true
+      t.integer :user_id, null: false, foreign_key: true
       t.integer :post_id, null: false, foreign_key: true
 
       t.timestamps
