+++ db/migrate/20220703054905_create_diaries.rb
@@ -1,7 +1,7 @@
 class CreateDiaries < ActiveRecord::Migration[6.1]
   def change
     create_table :diaries do |t|
-      t.bigint :user_id, null: false, foreign_key: true
+      t.integer :user_id, null: false, foreign_key: true
       t.integer :post_id, null: false, foreign_key: true
       t.string :title, null: false
       t.text :body, null: false
